feat: 도서 등록 API 구현 (Controller, Service, Repository)

+++ src/main/java/com/redhorse/deokhugam/domain/book/dto/response/BookDto.java
@@ -13,7 +13,7 @@ public record BookDto(
         LocalDate publishedDate,
         String isbn,
         String thumbnailUrl,
-        Long reviewCount,
+        int reviewCount,
         double rating,
         Instant createdAt,
         Instant updatedAt
